Initial commit
This gives an idea of where it's going but doesn't refresh the gravatars yet or add the stuff to the profile. But it's 5.17pm on a Friday afternoon and I am going on a proper holiday tomorrow so all of that will have to wait. I forgot the README though, maybe I'll do it on Github hey?